+++ vector/src/main/java/im/vector/app/features/roomprofile/polls/detail/ui/RoomPollDetailAction.kt
@@ -14,16 +14,10 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.roomprofile.polls.detail
+package im.vector.app.features.roomprofile.polls.detail.ui
 
-import im.vector.app.features.home.room.detail.timeline.item.PollOptionViewState
+import im.vector.app.core.platform.VectorViewModelAction
 
-data class RoomPollDetail(
-        val eventId: String,
-        val question: String,
-        val canVote: Boolean,
-        val votesStatusSummary: String,
-        val optionViewStates: List<PollOptionViewState>,
-        val hasBeenEdited: Boolean,
-        val isEnded: Boolean,
-)
+sealed interface RoomPollDetailAction : VectorViewModelAction {
+
+}
