+++ vector/src/main/java/im/vector/app/features/roomprofile/polls/detail/ui/RoomPollDetailViewEvent.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.roomprofile.polls.detail
+package im.vector.app.features.roomprofile.polls.detail.ui
 
 import im.vector.app.core.platform.VectorViewEvents
